Moving sharedID generation within sharedId module
